Move SoftClose translations into component
