+++ components/SoftClose/translations/de.js
