+++ components/SoftClose/translations/en.js
@@ -1,0 +1,8 @@
+const Translations = {
+	'soft-close-prompt': 'Do you want to support Wikipedia later?',
+	'soft-close-button-1': 'Yes',
+	'soft-close-button-2': 'Not now',
+	'soft-close-countdown-text': 'Automatically closes in <strong> {{seconds}} seconds</strong>.'
+};
+
+export default Translations;
